Use RFM to generate opacity table in NetCDF format (#14)
This PR should work although action fails. This is due to the fact that GitHub LFS only allows 1 GB bandwidth.
This PR is merged with risk. Later PR should use other methods than GitHub LFS.

+++ src/astro/Jupiter/jup_fletcher16_cirs.hpp
@@ -9,8 +9,8 @@
  */
 
 namespace Jupiter {
-extern int const fletcher16_cirs_npres;
-extern int const fletcher16_cirs_nglat;
+extern size_t const fletcher16_cirs_npres;
+extern size_t const fletcher16_cirs_nglat;
 extern double const fletcher16_cirs_glat[];
 extern double const fletcher16_cirs_pres[];
 extern double const fletcher16_cirs_temp[][120];
